DB/db_engine/requirements.txt
flask==2.1.2
flask-restx==0.5.1
gunicorn==20.1.0
requests==2.28.1
pyspark==3.2.1
py-libp2p==0.1.5
python-dotenv==0.19.2